Get delete working
As far as I can tell, Rails has deprecated link_to method: :delete.

Because of this, I have used button_to.

The downside to this is i cannot seem to get data-confirm working with that, either.

Added a project TODO to look into this further. For now, just going to rely on good flash messages and validations.